More test on checking the necessity of the override keyword on declaring functions
